bouh

+++ GITLAB_DOSS/day09/ex02/todo.js
@@ -1,12 +1,16 @@
 function PromptMessage()
 {
-	var entry = prompt("new to-do: ", "Enter text");
+	var entry = prompt("new to-do: ", "");
+	if (entry == "")
+		return;
 	console.log(entry);
 	var new_div = document.createElement("div");
 	new_div.setAttribute("onclick", "delete_node(this)");
 	new_div.appendChild(document.createTextNode(entry));
 	var mother = document.getElementById("ft_list");
 	mother.insertBefore(new_div, mother.childNodes[0]);
+	document.cookie = 'cookie1=' + entry + '; expires=Fri, 21 Feb 2020 00:00:00 UTC;';
+	console.log(document.cookie);
 }
 
 function delete_node(name)
